+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -18,8 +18,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 599;
-constexpr std::size_t translation_table_string_table_size = 82152;
+constexpr std::uint16_t translation_table_mapping_table_size = 600;
+constexpr std::size_t translation_table_string_table_size = 82202;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -504,6 +504,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "overload signature must match modifiers on this overload method"sv,
           "overloaded function '{0}' declared here"sv,
           "overloaded method is marked '{0}'"sv,
+          "override properties are not allowed in interfaces"sv,
           "parameter"sv,
           "parameter decorator must be before other modifiers"sv,
           "parameter decorators are not allowed in 'declare class'"sv,